ionbeam-cli/pyproject.toml
[project]
name = "ionbeam-cli"
version = "0.1.0"
description = "IonBeam CLI - Command line interface for managing ionbeam services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "aio-pika>=9.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "ionbeam-client",
]

[project.scripts]
ionbeam-cli = "ionbeam_cli.cli:main"

[tool.uv.sources]
ionbeam = { workspace = true }
ionbeam-client = { workspace = true }
acronet = { workspace = true }
eumetnet = { workspace = true }
ioncannon = { workspace = true }
meteogate = { workspace = true }
meteotracker = { workspace = true }
sensor-community = { workspace = true }
ecmwf = { workspace = true }
pygeoapi = { workspace = true }
legacy-api = { workspace = true }

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"